Integrated AppointmentViewModel
THE CURRENT BUILD CRASHES DUE TO A NULL OBJECT REFERENCE ON THE RETURNED
List<AppointmentModel> IN CALENDAR FRAGMENT.

I still need to add asynch functionality to AppointmentViewModel. That aside,
the appointment data is fetched and added to a List of <AppointmentModel>.
Appointment model is a helper class that stores the results of the
getApptByUsernameQuery.